app/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
httpx>=0.28.0
pytest>=8.0.0
pytrends>=4.9.2
